+++ MovieApp/MovieApp/Constant.swift
@@ -7,13 +7,6 @@
 
 import Foundation
 
-
-struct Colors {
-    static let background = "BgColor"
-    static let light_background = "LightBgColor"
-    static let placeholder = "PlaceholderTextColor"
-}
-
 struct identifiers {
     static let discover_collection_cell = "DiscoverCollectionCell"
     static let discover_collection_header = "DiscoverCollectionHeader"
@@ -26,26 +19,3 @@ struct identifiers {
     static let detail_trailer_collection_Cell = "DetailTrailerCollectionCell"
     static let detail_cast_collection_Cell = "DetailCastCollectionCell"
 }
-
-
-let genreCode = [
-    28 : "Action",
-    12 : "Adventure",
-    16 : "Animation",
-    35 : "Comedy",
-    80 : "Crime",
-    99 : "Documentary",
-    18 : "Drama",
-    10751 : "Family",
-    14 : "Fantasy",
-    36 : "History",
-    27 : "Horror",
-    10402 : "Music",
-    9648 : "Mystery",
-    10749 : "Romance",
-    878 : "Science Fiction",
-    10770 : "TV Movie",
-    53 : "Thriller",
-    10752 : "War",
-    37 : "Western"
-]
